+++ docs/src/markdown/about/changelog.md
@@ -2,6 +2,7 @@
 
 ## 4.2
 
+-   **NEW**: Provide new color spaces Okl~r~ab/OkL~r~Ch which use a modified lightness prediction.
 -   **NEW**: Add new matrix math functions that are specifically optimized for matrices and vectors of length 3 and
     leverage it in all appropriate places for a performance boost.
 -   **NEW**: Combine logic of `algebra` optimized vectorize functions and deprecate unnecessary function.